backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements-dev.in requirements.in
#
alembic==1.16.5
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
astroid==3.3.11
    # via pylint
bcrypt==4.3.0
    # via -r requirements.in
black==25.1.0
    # via -r requirements-dev.in
build==1.3.0
    # via pip-tools
click==8.2.1
    # via
    #   black
    #   pip-tools
    #   uvicorn
coverage[toml]==7.10.6
    # via pytest-cov
dill==0.4.0
    # via pylint
faker==37.6.0
    # via -r requirements-dev.in
fastapi==0.116.1
    # via
    #   -r requirements.in
    #   strawberry-graphql
flake8==7.3.0
    # via -r requirements-dev.in
graphql-core==3.2.6
    # via strawberry-graphql
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via
    #   -r requirements-dev.in
    #   pylint
lia-web==0.2.3
    # via strawberry-graphql
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy==1.17.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   build
    #   pytest
    #   strawberry-graphql
pathspec==0.12.1
    # via
    #   black
    #   mypy
pip-tools==7.5.0
    # via -r requirements-dev.in
platformdirs==4.4.0
    # via
    #   black
    #   pylint
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg[binary]==3.2.9
    # via -r requirements.in
psycopg-binary==3.2.9
    # via psycopg
pycodestyle==2.14.0
    # via flake8
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via -r requirements.in
pylint==3.3.8
    # via -r requirements-dev.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.1.0
    # via -r requirements-dev.in
pytest-cov==6.2.1
    # via -r requirements-dev.in
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   -r requirements.in
    #   strawberry-graphql
pyyaml==6.0.2
    # via uvicorn
redis==6.4.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.47.3
    # via fastapi
strawberry-graphql[fastapi]==0.281.0
    # via -r requirements.in
tomlkit==0.13.3
    # via pylint
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   lia-web
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   strawberry-graphql
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via faker
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
